+++ app/src/main/java/com/github/bumblebee202111/minusonecloudmusic/data/repository/PlaylistRepository.kt
@@ -1,6 +1,5 @@
 package com.github.bumblebee202111.minusonecloudmusic.data.repository
 
-import androidx.core.net.toUri
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
@@ -9,10 +8,7 @@ import androidx.room.withTransaction
 import com.github.bumblebee202111.minusonecloudmusic.coroutines.ApplicationScope
 import com.github.bumblebee202111.minusonecloudmusic.data.Result
 import com.github.bumblebee202111.minusonecloudmusic.data.database.AppDatabase
-import com.github.bumblebee202111.minusonecloudmusic.data.database.model.entity.LocalSongEntity
 import com.github.bumblebee202111.minusonecloudmusic.data.database.model.entity.PlayerPlaylistSongEntity
-import com.github.bumblebee202111.minusonecloudmusic.data.database.model.entity.RemoteSongEntity
-import com.github.bumblebee202111.minusonecloudmusic.data.database.model.entity.asExternalModel
 import com.github.bumblebee202111.minusonecloudmusic.data.database.model.view.GenericSongView
 import com.github.bumblebee202111.minusonecloudmusic.data.database.model.view.asExternalModel
 import com.github.bumblebee202111.minusonecloudmusic.data.model.AbstractRemoteSong
@@ -23,14 +19,12 @@ import com.github.bumblebee202111.minusonecloudmusic.data.model.PlaylistDetail
 import com.github.bumblebee202111.minusonecloudmusic.data.model.RemoteSong
 import com.github.bumblebee202111.minusonecloudmusic.data.model.SongIdAndVersion
 import com.github.bumblebee202111.minusonecloudmusic.data.model.asEntity
-import com.github.bumblebee202111.minusonecloudmusic.data.model.toAudioId
 import com.github.bumblebee202111.minusonecloudmusic.data.network.NetworkDataSource
 import com.github.bumblebee202111.minusonecloudmusic.data.network.model.ApiResult
 import com.github.bumblebee202111.minusonecloudmusic.data.network.model.music.NetworkBillboardGroup
 import com.github.bumblebee202111.minusonecloudmusic.data.network.model.music.SongDetailsApiModel
 import com.github.bumblebee202111.minusonecloudmusic.data.network.model.music.asExternalModel
 import com.github.bumblebee202111.minusonecloudmusic.data.network.requestparam.CParamSongInfo
-import com.github.bumblebee202111.minusonecloudmusic.player.isLocalSongUri
 import com.github.bumblebee202111.minusonecloudmusic.player.mediaIdToIsLocalAndSongId
 import com.squareup.moshi.JsonAdapter
 import kotlinx.coroutines.CoroutineScope
@@ -160,7 +154,7 @@ class PlaylistRepository @Inject constructor(
         }
     }
 
-    suspend fun getPlayerPlaylistSongPosition(mediaId: String): Int {
+    suspend fun getPlayerPlaylistSongPosition(mediaId: String): Int? {
         return playerDao.getPlaylistSongPosition(mediaId)
     }