+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Change Log
 ## Version 1.0.0-dev05 (baseline: NCM 9.0.6x)
 
 - Fix crashes when you are trying to play playlists
-- Fix size of navigationIcons Bumblebee202111
+- Fix size of navigationIcons
 - Fix app crash when the guest user is trying to open cloud disk
 - Fix logged-in user getting altered by guest on refresh
 - Update dependencies